WIP reading dep file
